+++ src/allianceutils/middleware/__init__.py
@@ -1,3 +1,4 @@
+from .current_request import CurrentRequestMiddleware
 from .current_user import CurrentUserMiddleware
 from .http_auth import HttpAuthMiddleware
 from .query_count import QueryCountMiddleware
@@ -6,4 +7,5 @@ __all__ = [
     'HttpAuthMiddleware',
     'CurrentUserMiddleware',
     'QueryCountMiddleware',
+    'CurrentRequestMiddleware',
 ]